Jwt localstorage

+++ src/main/resources/static/js/usuarios.js
@@ -7,10 +7,7 @@ $(document).ready(function() {
 async function  cargarUsuarios(){
   const request = await fetch('api/usuarios', {
     method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json'
-    },
+    headers: getHeaders()
    
   });
   const usuarios = await request.json();
@@ -41,14 +38,23 @@ async function  cargarUsuarios(){
   document.querySelector('#usuarios tbody').outerHTML = listHTML;
 }
 
+function getHeaders(){
+  return {
+    'Accept': 'application/json',
+    'Content-Type': 'application/json',
+    'Authorization':localStorage.token
+  };
+}
+
 async function eliminarUsuario(id){
+  if (!confirm('Desa elminar el user')){
+    return;
+  }
+
   const request = await fetch('api/usuarios/'+id, {
     method: 'DELETE',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json'
-    },
+    headers: getHeaders()
 
   });
-
+  location.reload()
 }
